+++ src/commands.rs
@@ -1,8 +1,10 @@
+pub mod exec;
 pub mod ping;
 pub mod push;
 pub mod repl;
 pub mod reset;
 
+pub use exec::exec;
 pub use ping::ping;
 pub use push::push;
 pub use repl::run_repl;